server/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.6.2
black==21.6b0
click==8.0.1
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
isort==5.9.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
pylint==2.9.3
pymongo==3.12.0
pytz==2021.1
regex==2021.7.6
sqlparse==0.2.4
toml==0.10.2
wrapt==1.12.1
